Move the swarm functionality to the p2p layer
