+++ transport/p2p/src/constants.rs
@@ -9,6 +9,6 @@ pub(crate) const HOPR_TICKET_AGGREGATION_PROTOCOL_V_0_1_0: &str = "/hopr/ticket-
 
 // Swarm configuration
 /// The maximum number of concurrently dialed (outbound) peers.
-pub const HOPR_SWARM_CONCURRENTLY_DIALED_PEER_COUNT: u8 = 255;
+pub(crate) const HOPR_SWARM_CONCURRENTLY_DIALED_PEER_COUNT: u8 = 255;
 /// The maximum number of concurrently negotiating inbound peers.
-pub const HOPR_SWARM_CONCURRENTLY_NEGOTIATING_INBOUND_PEER_COUNT: usize = 255;
+pub(crate) const HOPR_SWARM_CONCURRENTLY_NEGOTIATING_INBOUND_PEER_COUNT: usize = 255;
